+++ app/ante/cosmos/authz_test.go
@@ -403,16 +403,22 @@ func (suite *AnteTestSuite) TestRejectMsgsInAuthz() {
 			if tc.isEIP712 {
 				coinAmount := sdk.NewCoin(evmtypes.DefaultEVMDenom, sdk.NewInt(20))
 				fees := sdk.NewCoins(coinAmount)
+				cosmosTxArgs := utiltx.CosmosTxArgs{
+					TxCfg:   suite.clientCtx.TxConfig,
+					Priv:    suite.priv,
+					ChainID: suite.ctx.ChainID(),
+					Gas:     200000,
+					Fees:    fees,
+					Msgs:    tc.msgs,
+				}
+
 				tx, err = utiltx.CreateEIP712CosmosTx(
 					suite.ctx,
 					suite.app,
-					utiltx.CosmosTxArgs{
-						TxCfg:   suite.clientCtx.TxConfig,
-						Priv:    suite.priv,
-						ChainID: suite.ctx.ChainID(),
-						Gas:     200000,
-						Fees:    fees,
-						Msgs:    tc.msgs,
+					utiltx.EIP712TxArgs{
+						CosmosTxArgs:       cosmosTxArgs,
+						UseLegacyExtension: true,
+						UseLegacyTypedData: true,
 					},
 				)
 			} else {